requirements.txt
requests
oauthlib
openai
feedparser  
typer
python-dotenv
python-dateutil
PyYAML